--- Assets/Sandbox/Dariia/Casket/Materials.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 55d9f7a3a4438174f93fdf010fe60455
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
