--- Assets/Sandbox/Dariia/Fence/Material.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 16460ca79acd5ca4b92de06700cba40a
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
