--- Assets/Sandbox/Dariia/Fence/Prefabs/FenceParts.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 149ee0004e49bfd499a5a338c7598d15
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
