--- Assets/Sandbox/Dariia/Fence/Scenes.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 2946c5308bbf5a446a8e77ef2af3b19b
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
