--- Assets/Sandbox/Dariia/key/Materials.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 9e522e75570e6b446b5f90a8fab188b6
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
