--- Assets/Sandbox/Dariia/sounds.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: c1de80a422879cb4e96a17df7e3105db
-folderAsset: yes
-timeCreated: 1516799460
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
