+++ Assets/Sandbox/Tests/KeyLockedDoor.meta
