+++ Assets/Sandbox/Tests/KeyLockedDoor/Materials.meta
