+++ Assets/Sandbox/Tests/KeyLockedDoor/Materials/Gold.mat
