+++ Assets/Sandbox/Tests/KeyLockedDoor/Prefabs.meta
