+++ Assets/Sandbox/Tests/KeyLockedDoor/Prefabs/Door.prefab.meta
