+++ Assets/Sandbox/Tests/KeyLockedDoor/Prefabs/Key.prefab
