+++ Assets/Sandbox/Tests/KeyLockedDoor/Scenes/DoorWithKeyTest.unity.meta
