+++ Assets/Sandbox/Tests/KeyLockedDoor/Scripts.meta
