+++ Assets/Sandbox/Tests/KeyLockedDoor/Scripts/DoorKey.cs.meta
