+++ Assets/Sandbox/Tests/KeyLockedDoor/Scripts/KeyLockedDoor.cs.meta
