--- Assets/Sandbox/Tests/Template/Prefabs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 5f2600767eeaa5f4fb28e3752e605856
-folderAsset: yes
-timeCreated: 1515714786
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
